% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbhy.optr-package.R
\docType{package}
\name{pbhy.optr-package}
\alias{pbhy.optr}
\alias{pbhy.optr-package}
\title{pbhy.optr: General-Use Population-Based Hybrid Optimization}
\description{
todo.
}
\author{
\strong{Maintainer}: Ricardo Semião \email{ricardo.semiao@outlook.com}

}
\keyword{internal}
